requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
google-cloud-vision==3.4.4
google-cloud-storage==2.13.0
google-cloud-aiplatform==1.70.0
pinecone==5.4.2
PyPDF2==3.0.1
pillow==10.1.0
python-jose==3.3.0
httpx==0.25.2
neo4j==5.14.0